Fix fixed time issue The problem was caused when rounding time_t to float. Since there is no use for the decimal point, changing timestamp to unsigned int fixes the issue.

+++ server/src/Communication/CommunicationManager.h
@@ -14,7 +14,7 @@ public:
     CommunicationManager(char *ipAddress, unsigned int port);
 
     Packet listen();
-    Packet putTimestamp(Packet packet, float timestamp);
+    Packet putTimestamp(Packet packet, unsigned int timestamp);
 
     void handleConnection(sockaddr_in clientAddress);
 };
