+++ server/src/Server.cpp
@@ -6,6 +6,6 @@ Packet Server::listen(){
     return this->communicationManager.listen();
 }
 
-Packet Server::putTimestamp(Packet packet, float timestamp){
+Packet Server::putTimestamp(Packet packet, unsigned int timestamp){
     return this->communicationManager.putTimestamp(packet, timestamp);
 }
